AppSettings: revert default environment from Local to Development and give to Local precedence over anything else
The application fails to build from a fresh clone. This is due to the fact the appsettings.Local.json is mandatory and not integrated in the repository.
Conceptually the appsettings.Local.json should be optional and give the possiblity to override appsettings.Development.json's default settings.
Default connection string for Catalog has been set to "(LocalDB)\\MSSQLLocalDB; Database=Catalog; Integrated Security=true;".