Domain: add support to configure a trainer's social networks (#15)
This commit introduces an enumeration of supported Social network.
It also adds a new many to many table between trainers and social networks enumeration.

This commit introduce a new enumeration that represents a Social Network and also a new many to many table linking a trainer and a social network.
The model of this table contains a string property to optionally specify the trainer's URL to a given social network.