Layouts: use the Smart.Design.Razor library static css files
So far the linked style sheet is https://design.smart.coop/css/main.css, let's use the library one to avoid "surprises" in the future.
Also to mae this possible, we had to change a bit the start up configuration because by default static files are only resolved on development builds.