Use the framework middleware that handles redirection on error http status codes
We were using a simple minimal middleware to redirect 404 errors to /404.
However if a page returns an instance of NotFoundResult than this middleware doesn't handle properly the redirection.
Microsoft's one works pretty well, so why not use it in place of our ?
Note: the middleware checks for responses with status codes between 400 and 599 that do not have a body.