Prevent internal error when editing a training
Because the topics were not correctly loaded, they were never cleared correctly during the save.
This resulted into a duplication primary sql error.
Making them tracked corrects the problem.
Also made sure to include the topic's ids in the view model of the page otherwise the existing topic are not checked.