Fix a JS issue on loading of almost every pages
There was  a onload event logic applied globally to every pages but that logic is related only to creation and edition pages.
Therefore opening any pages but those two would result in this:
'Cannot read properties of null (reading 'addEventListener')'.
This commit provides a quick and dirty fix by returning in nothing is found.