Add CreatedBy and LastModifiedBy properties to Entity base class
This commits makes auditable the CreateBy, CreatedAt, LastModifiedBy and LastModifiedAt properties.
CatalogContext will ensure that those properties are correctly set upon saving.
The context will not update modified properties when inserting an entity nor will update created properties when updating an entity.
The CreatedBy and LastMofifiedBy properties are set accordingly to the Id of the current user.
CreatedAt and LastModifiedAt properties are set at DateTime.UtcNow if not ignored.