LoggingPipelineBehavior: don't log Trainer nor Training domain models that transit through MediatR
Due to cyclic references between properties of Trainer and Training (they are aggregate after all) the serialization writes very very large json in the output.
This commit adds two JsonConverters on Training and Trainer to convert them to their primary key value rather the whole object value.
Another issue with serializing domain models is that it will trigger the evaluation of every unloaded properties, which can result in an unnecessary SQL statement.