Update description form with image upload action (#249)
This PR relates to description form of the trainer being "deleted" while uploading an image (it's not per-se deleted, it's just if no call has been done to save the form and that an image upload refresh the page, clearly the text of the form will be lost).
The PR simply create a form for both the upload of an image and the update of the description fields. The Save button keeps it's previous use of only saving the descriptions fields (and not uploading an image).